Clean and fix historical interpolation for new SSP GDP
